+++ pytket/tests/placement_test.py
@@ -113,8 +113,6 @@ def test_placement_config() -> None:
     test_pl = GraphPlacement(test_architecture)
     c = Circuit(4).CX(0, 1).CX(1, 2).CX(2, 3)
     pm = test_pl.get_placement_map(c)
-    with pytest.deprecated_call():
-        test_pl.modify_config()
     assert test_pl.get_placement_map(c) == pm